🧪 [automation]: Setup generate-docs script for GH workflow

+++ scripts/shared/spinner.sh
@@ -9,7 +9,7 @@ EMPTY=""
 LOADER_SEQUENCE_1=("⊏" "⊓" "⊐" "⊔")
 LOADER_SEQUENCE_2=("⣾" "⣽" "⣻" "⢿" "⡿" "⣟" "⣯" "⣷")
 LOADER_SEQUENCE_3=("𓃉𓃉𓃉" "𓃉𓃉∘" "𓃉∘°" "∘°∘" "°∘𓃉" "∘𓃉𓃉")
-LOADER_SEQUENCE_4=("|" "/" "-" "\\")
+LOADER_SEQUENCE_4=("/" "-" "\\" "|")
 
 LOADER_SEQUENCE=("${LOADER_SEQUENCE_4[@]}")
 
@@ -25,14 +25,20 @@ spinner() {
 }
 
 start() {
-    SPINNER_MSG_LEFT="${1:-$EMPTY}"
-    SPINNER_MSG_RIGHT="${2:-$EMPTY}"
+  SPINNER_MSG_LEFT="${1:-$EMPTY}"
+  SPINNER_MSG_RIGHT="${2:-$EMPTY}"
+  if [ "$SPINNER_DISABLED" == "true" ]; then
+    echo -e "$SPINNER_MSG_LEFT${LOADER_SEQUENCE[0]}$SPINNER_MSG_RIGHT"
+  else
     echo ""
     spinner "$SPINNER_MSG_LEFT" "$SPINNER_MSG_RIGHT" & SPINNER_PID=$!
+  fi
 }
 
 stop() {
+  if [ "$SPINNER_DISABLED" != "true" ]; then
     FINISH_REPLACER="${1:-$INITIAL_SPINNER_STATE}"
     kill $SPINNER_PID 
     echo -ne "\r$SPINNER_MSG_LEFT$FINISH_REPLACER$SPINNER_MSG_RIGHT"
+  fi
 }
